deck-service/functions/AddCardToDeck/requirements.txt
aws_xray_sdk
boto3
python-jose
requests
responses
